requirements.txt
numpy==1.13.3
Pillow==4.3.0